Enable prevent_unsafe_narrowing for //sandbox/win:sandbox
Updates remaining smaller sites to compile under
prevent_unsafe_narrowing and turns on for //sandbox/win:sandbox.

* NTSTATUS_ codes added to nt_internals.h where their definition
  would otherwise conflict with DWORD STATUS_ codes.
* casting of some pointer arithmetic results to size_t.
* size_t or DWORD for counts/nums where appropriate.

Bug: 1385207
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4029619
Commit-Queue: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1073717}

+++ sandbox/win/src/handle_closer.cc
@@ -10,6 +10,7 @@
 
 #include "base/check_op.h"
 #include "base/memory/free_deleter.h"
+#include "base/numerics/checked_math.h"
 #include "base/ranges/algorithm.h"
 #include "sandbox/win/src/win_utils.h"
 
@@ -141,8 +142,8 @@ bool HandleCloser::SetupHandleList(void* buffer, size_t buffer_bytes) {
     i->first.copy(output, i->first.size());
     *(output += i->first.size()) = L'\0';
     output++;
-    list_entry->offset_to_names =
-        reinterpret_cast<char*>(output) - reinterpret_cast<char*>(list_entry);
+    list_entry->offset_to_names = base::checked_cast<size_t>(
+        reinterpret_cast<char*>(output) - reinterpret_cast<char*>(list_entry));
     list_entry->name_count = i->second.size();
 
     // Copy the handle names.
@@ -153,8 +154,8 @@ bool HandleCloser::SetupHandleList(void* buffer, size_t buffer_bytes) {
 
     // Round up to the nearest multiple of sizeof(size_t).
     output = RoundUpToWordSize(output);
-    list_entry->record_bytes =
-        reinterpret_cast<char*>(output) - reinterpret_cast<char*>(list_entry);
+    list_entry->record_bytes = static_cast<size_t>(
+        reinterpret_cast<char*>(output) - reinterpret_cast<char*>(list_entry));
   }
 
   DCHECK_EQ(reinterpret_cast<size_t>(output), reinterpret_cast<size_t>(end));
